Websocket PoC: server now pushes to clients (#6)
Gone is the relaying of messages from any client to the rest of the
pool, and now Client messages are just received and processed by the
Server, while the Server Stdin is relayed to all the peers.
This closes the PoC in a successful note and opens the door for
architectural discussions on how to proceed with the development.

Changelog:
* encapsulate stdin reading into its own fn
* refactor client message writing to server
* remove client message relay, now server broadcast from its own stdin