add first no-std client prototype (#79)
This prototype is a proof of concept for a client that runs in a rp2040.

It connects to the local WiFi, connects to the local Lamarrs server, registers and subscribe to the color service.

It also has a small OLED SSD1306 screen connected too, and reports the received messages to that device. The screen simulates any other message processor, like a DMX driver, LED controller, etc.

The prototype was successful, albeit much more work is needed to increment reliability, extensibility and usability.