rework utils: remove midi mirrored structs
These structs were a hack to add serde to the classes used to send MIDI payloads to the web ui.
Since oxysynth has been updated and the web ui is currently not in development anymore, removing the hack seems the right thing to do in oder the bring some order to this chaos.