rework utils: new client-server messages
A complete rework of the messages between Server and Clients. These messages are no-std compatible, enabling us to use them in the rp-client too.
ExchangeMessage wraps Actions to be executed by the receptor, while Data wraps the information that will be required by the receptor to execute the requested action.

Dividing ExchangeMessages between Client msg and Server msg may bring dividens in the future, but by now, we are fine mixing them all.