Major refactor (#83)
* add rp-client targets to .gitignore
* update cargo.lock
* update .gitignore
* rework utils: remove midi mirrored structs
These structs were a hack to add serde to the classes used to send MIDI payloads to the web ui.
Since oxysynth has been updated and the web ui is currently not in development any more, removing the hack seems the right thing to do in order to bring some order to this chaos.

* rework utils: remove old, refactored modules
The old modules division didn´t make a lot of sense. All of these are being replaced by newer, more domain driven data structures and enums, making these obsolete.

* rework utils: new client-server messages
A complete rework of the messages between Server and Clients. These messages are no-std compatible, enabling us to use them in the rp-client too.
ExchangeMessage wraps Actions to be executed by the receptor, while Data wraps the information that will be required by the receptor to execute the requested action.

Dividing ExchangeMessages between Client msg and Server msg may bring dividens in the future, but by now, we are fine mixing them all.

* rework-utils: new Orchestrator messages
Orchestrator is still WIP, and by now only makes request to the server, writing to MQTT.

* rework-utils: bump versions, remove no-std libs
* rework-utils: make concrete data types no-std compatible
Here are all the Structs with the concrete data that is wrapped in the onion of Enums used to send messages between Server and Client.

* rework-orchestrator: major refactor, drop MIDI
Since most of the messages and payloads have changed, orchestrator had to be adapted to all these changes to keep it working. This interface will be reworked soon heavily soon, removing MQTT as a carrier and creating a direct communication with the Server.

* rework-orchestrator: bump and add dependencies
* rework-client: major redevelopment
Basically, this is an entire brand-new development. The new client can playback files that have been previously loaded to the device that is running it. This is an `std` client, for a `no-std` use `rp-client`.

* add cross files to compile to RPi 3B - WIP
This is not working as `rodio` keeps trying to link local alsa during building instead of the container one. Eventually will be fixed, by now I'm compiling on target.

* rework-utils: reword messages to improve readeability
* rework-client: adapt to changes in utils
* rework-orchestrator: adapt to changes in utils
* rework-server: rename and rework connection listener
* rework-server: rename and refactor ws handler
* rework-server: refactor, fix eyre.
* rework-server: create trait for services
Most of the Server services do almost the same so far, so it makes sense they have a base implementation of the basic services and, if they need custom implementations of them, branch in the service itself.

* rework-server: remove obsolete services
* rework-server: create new services
* rework-server: WIP, add new tests (broken)
* rework-server: adapt MQTT service to changes in utils and server
* rework-server: bump deps, add heapless and eyre
* rework-web: WIP, rework in progress, still broken