add binary payload processing, encapsulate json decoding in handle_ws_message
The previous implementation was decoding the json way later in the "message peeling" handlers. This bleed reponsibilities from one function to the other, increasing the coupling between message types.
When we tried to implement the Binary processing, this became evident. Changing the arguments for the Message processors to be Exchange Messages made clear that handle_ws_message should be the one deserialising the Text payload to the concrete ExchangeMessage.