Integrate `lamarrs-utils` to `rp-client`
* rework utils: new client-server messages
A complete rework of the messages between Server and Clients. These messages are no-std compatible, enabling us to use them in the rp-client too.
ExchangeMessage wraps Actions to be executed by the receptor, while Data wraps the information that will be required by the receptor to execute the requested action.

Dividing ExchangeMessages between Client msg and Server msg may bring dividens in the future, but by now, we are fine mixing them all.

* add Display to lamarrs-util structs
Quite convenient for error displaying

* add lamarr-utils and replace serde-json-core for postcard
* add rp-client to the workspace list
* cargo fmt
* add binary payload processing, encapsulate json decoding in handle_ws_message
The previous implementation was decoding the json way later in the "message peeling" handlers. This bleed reponsibilities from one function to the other, increasing the coupling between message types.
When we tried to implement the Binary processing, this became evident. Changing the arguments for the Message processors to be Exchange Messages made clear that handle_ws_message should be the one deserialising the Text payload to the concrete ExchangeMessage.

* cargo fmt
* add postcard dep to the server
* remove ws_task from main, will be in its own module
* rename module
* add rp-client targets to .gitignore
* fix cargo.lock