add postcard deps for the std client
Since we are introducing a watchdog, we will need to default all comms to binary until the client sends its first message declaring the kind of wire to use. Clients can use Text or Binary.
The no-std clients don't play nice with text payloads, so they will keep using binary. But for other clients, using json is the more maintainable way to go, mainly for debugging if something goes wrong.