adjust service traits to changes needed to work with watchdog and reconnect
Upsert was mixing actions. We didn´t always wanted to update or insert if not existant, because during reconnection we only want to upate IF the client uuid was already registered.

This fixes the issues, improves reliability of the services, improves logging and also introduces cleanup management in case the watchdog kicks in.