Create RP Client trigger
* add postcard deps for the std client
Since we are introducing a watchdog, we will need to default all comms to binary until the client sends its first message declaring the kind of wire to use. Clients can use Text or Binary.
The no-std clients don't play nice with text payloads, so they will keep using binary. But for other clients, using json is the more maintainable way to go, mainly for debugging if something goes wrong.

* add display for ClientIdAndLocation
* add defmt dep to utils
* rename ExchangeMessage, add Next, RetriggerScene and Heartbeats
The previous name for Update was confusing. This new name makes more sense.
I still think we need to split these messages better.

* add Binary message processing for std Clients
* introduce oled module for rp-client
* move oled control to another module in rp-client
* add external gpio input processor in rp-client
This allows connecting a button to send fixed messages to the server. Useful for a trigger to move scenes forward.

* adapt ws handler in rp-client to use binary and not text
* adapt orchestrator sender to new message name
* major refactor of rp-client session holder
These changes introduce eternal loop to keep the connection alive, reconnection mechanism, better error handling, better logging, handling of incoming messages and watchdog heartbeat requests management.

* add client manager wire poliglot processing, watchdog and cleanup mechanism in server
* adjust service traits to changes needed to work with watchdog and reconnect
Upsert was mixing actions. We didn´t always wanted to update or insert if not existant, because during reconnection we only want to upate IF the client uuid was already registered.

This fixes the issues, improves reliability of the services, improves logging and also introduces cleanup management in case the watchdog kicks in.

* adjust to changes in the Messages naming
* fix trigger and retrigger logic
* add impl for Action to convert to str
* report watchdog in rp-client oled in the lamarrs status line
* add deps to server to create the sequencer, add clap to add a runner cli
* add sequence parser for yaml messages to set the sequencer
* add sequencer service in the server
* add sequencer to main, add clap cli for server
* client factory now inject sequencer sender and params
* add server handling for next scene and retrigger scene commands
* adjust orchestrator to latest changing in message naming
* add example of sequencer config