Multiple, importan changes, add midi, improve client robustness
I never do this kind of massive commits without any explanation of what's going on, but we have a show in 3 days, I slept only 4hs and there are feces and piss from the neighbors coming down my kitchen sink because somebody clogged the general sewage drainage and we live in the ground floor. I need to commit all of this fast while we continue adding support for more things.

Basically, the Client refactor added similar mechanisms to the no-std Client, with some kinda clunky error handling that we will refactor later.
The Client now doesn´t block anymore on sound reproduction and detaches from it in some fire-and-forget fashion.

Major add on is the Midi Service: now we an send Midi commands via USB Midi in a std Client. This will be used to control a voice pedal mainly.
Also, replace the serde-yaml for serde-yml that seems to be its evolution. It still can´t parse nested enums.

Again, sorry me from the future or however is watching this, I feel very ashamed by doing it, but a man covered in piss and shit and without water at its place can be forgiven of some sins.